pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool]

[tool.dephell]

[tool.dephell.main]

[tool.dephell.main.from]
format = "poetry"
path = "pyproject.toml"

[tool.dephell.main.to]
format = "pip"
path = "requirements.txt"

[tool.poetry]
authors = ["Tony Brown <tb148@tuta.io>"]
description = "A discord bot that works."
license = "AGPL-3.0-or-later"
maintainers = ["Tony Brown <tb148@tuta.io>"]
name = "grogarium"
readme = "README.rst"
repository = "https://github.com/tb148/grogarium/"
version = "0.3.0"

[tool.poetry.dependencies]
googletrans = "^4.0.0rc1"
python = "^3.8"
sympy = "^1.8"
toml = "^0.10.2"
chess = "^1.6.1"
google-trans-new = "^1.1.9"
"discord.py" = "^1.5.1"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
coverage = "^5.3"
docutils = "^0.17"
dpytest = "^0.5.3"
sphinx = "^4.1.2"
poetry = "^1.1.8"
dephell = "^0.8.3"
jedi = "^0.18.0"
autopep8 = "^1.5.7"
pylama = "^7.7.1"

[tool.autopep8]
max_line_length = 80
in-place = true
recursive = true
aggressive = 3
